Move config to models
